XIVY-6717 Hide connect quick action
Hide connect quick action on elements which only allow one connection
