+++ editor/src/quick-action/edge/model.ts
@@ -1,0 +1,7 @@
+import { SModelElement } from '@eclipse-glsp/client';
+
+export const multipleOutgoingEdgesFeature = Symbol('multipleOutgoingEdgesFeature');
+
+export function isMultipleOutgoingEdgesFeature(element: SModelElement): element is SModelElement {
+  return element.hasFeature(multipleOutgoingEdgesFeature);
+}
